refactor: update project for sprint 10
