+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -14,30 +14,30 @@ import java.util.List;
 @RequestMapping("/users")
 @Slf4j
 public class UserController {
-    private UserService userService;
+    private UserStorage userStorage;
 
     @Autowired
-    public UserController(UserService userService) {
-        this.userService = userService;
+    public UserController(UserStorage userStorage) {
+        this.userStorage = userStorage;
     }
 
     @GetMapping
     public List<User> getAllUsers() {
         log.debug("Сделан запрос на получение списка всех пользователей");
-        return userService.getAllUsers();
+        return userStorage.getAllUsers();
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public User createUser(@Valid @RequestBody User user) {
         log.debug("Сделан запрос на добавление нового пользователя");
-        return userService.createUser(user);
+        return userStorage.createUser(user);
     }
 
     @PutMapping
     @ResponseStatus(HttpStatus.OK)
     public User updateUser(@Valid @RequestBody User user) {
         log.debug("Сделан запрос на обновление пользователя с ID={}", user.getId());
-        return userService.updateUser(user);
+        return userStorage.updateUser(user);
     }
 }
