+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,0 +1,13 @@
+package ru.yandex.practicum.filmorate.storage.user;
+
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.List;
+
+public interface UserStorage {
+    List<User> getAllUsers();
+
+    User createUser(User user);
+
+    User updateUser(User user);
+}
