+++ src/main/java/ru/yandex/practicum/filmorate/utils/IdGenerator.java
@@ -1,0 +1,16 @@
+package ru.yandex.practicum.filmorate.utils;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class IdGenerator {
+    private Integer idGenerator = 1;
+
+    public Integer getIdGenerator() {
+        return idGenerator++;
+    }
+
+    public void resetIdGenerator() {
+        this.idGenerator = 1;
+    }
+}
